Added outlined text shaders
Small text shader needs additional work. Testing has indicated that it It needs to be
separated into multiple passes for proper rendering.

Need to see how to make the stroke-only fatter on demand, that
or make the shader optimized for small text a bit skinnier on demand.

Clearly, at small sizes, it is best done in two passes, like how we observe Adobe
software doing it.